ENH: Add additional environment to `tox` to run notebooks
Add an additional environment to `tox` to run notebooks.

Add a new table to `pyproject.toml` to host the `jupyter`, `nbclient`
and `nbmake` dependencies so that the notebooks can be run.

Install the necessary TeX packages in the test CI so that LaTeX
decoration in plots can be employed.

Install the necessary additional Python packages used across the
notebooks.

Add a helper Python script to actually run the notebooks since `tox`
does not expand wildcards currently.