REF: Refactor filters into `filtering`
Refactor code blocks that perform filtering operations within models
into the `filtering` module so that the model and filter concepts are
separated, supporting more cleanly pipelining models with filters within
the new `Estimator` philosophy.

Add the percentile-based detrending feature that was temporarily removed
in commit 7322e93.

Take advantage of the commit to rename the `th_low` and `th_high`
attributes of the `AverageDWIModel` to `atol_low` and `atol_high`.

Co-authored-by: Oscar Esteban <code@oscaresteban.es>