REF: Refactor filters into `filtering`
Refactor code blocks that perform filtering operations within models
into the `filtering` module so that the model and filter concepts are
separated, supporting more cleanly pipelining models with filters within
the new `Estimator` philosophy.

Add the percentile-based detrending feature that was temporarily removed
in commit 7322e93.

Take advantage of the commit to rename the `th_low` and `th_high`
attributes of the `AverageDWIModel` to `atol_low` and `atol_high`.

Fix the mask treatment in the robust minmax normalization computation
function: when the original mask value is not `None`, the `data[mask]`
indexing results in an error since the mask is a 1D array whose
dimension does not match the indices along dimension 0 of the data
array.

Fixes:
```
>       reshaped_data = data.reshape((-1, data.shape[-1])) if mask is None else data[mask]
E       IndexError: boolean index did not match indexed array along dimension 0; dimension is 32 but corresponding boolean dimension is 5
```

Co-authored-by: Oscar Esteban <code@oscaresteban.es>