FIX: Add null gradient values to DWI data serialization if appropriate
Add null gradient values to DWI data serialization if appropriate.

When the flag `insert_b0` is `True` the previous implementation
prepended the `bzero` attribute data to the `DWI` instance `dataobj`,
but it was not adding the corresponding null gradient value to the
bval/bvec file pair. This patch set prepends a null gradient to the
bval/bvec file pair.

Parameterize the `test_load` testing function to check both cases (i.e.
`insert_b0` `False` and `True`).

Add additional checks to ensure that the write/read round-trip works as
expected, and that the `bvals` and `bvecs` attributes have the expected
values.

Take advantage of the commit to make it clear in the `DWI.to_nifti`
method that it assumes that the `bzero` attribute value is never `None`.