enh: allow "locking" of models with first fit
This allows having "frozen" models that are fit only once before
entering the leave-one-out loop.

These models use all the data in the fitting and then return always the
same "prediction".

This feature was lost with the refactor of the estimator. By moving it
into the models, we can use them in a more flexible way.