TST: Do not remove fMRI BOLD notebook from list of discovered notebooks
Do not remove fMRI BOLD realignment notebook from list of discovered
notebooks: this will allow running the notebook at the time scheduled by
the `notebooks.yml` GHA workflow file.

Download `SynthStrip` model for skull stripping. `SynthStrip` is used by
the fMRI BOLD realignment notebook.

Fixes:
```
FileNotFoundError                         Traceback (most recent call last)
Cell In[4], line 19
     14 if not bmask_path.exists():
     15     bmsk_results = SynthStrip(
     16         in_file=str(avg_path),
     17         use_gpu=True,
     18     ).run(cwd=str(WORKDIR))
---> 19     copy(bmsk_results.outputs.out_mask,bmask_path)

(...)

FileNotFoundError: [Errno 2] No such file or directory:
 '/home/runner/tmp/nifreezedev/ismrm25/sub-01_task-mixedgamblestask_run-03_boldref_desc-brain_mask.nii.gz'
```

raised for example at:
https://github.com/jhlegarreta/nifreeze/actions/runs/14897027210/job/41841440531#step:11:102

Modify the call to the `SynthStrip` workflow in the notebook so that the
model is loaded from a file that is provided as an argument.

Install AFNI in the GHA workflow.

Check if `CUDA` is available in the GHA workflow file and set an
environment variable to run the `SynthStrip` process on CPU if not
available.