enh: improved handling of the *b=0* and a max *b* in dMRI datasets
**Maximum *b*-value**: Models like DTI and DKI will now have a ``b_max``
configuration that drops *b*-values above it (defaults: 1500 and 2500,
respectively).
This configuration can be modified at the time of model instantiation.
One side effect is that, in multi-shell datasets, the fitting of DTI
and DKI is normally sped up at no cost (even improving) because fitting
them with exceedingly high b-values is counterproductive.

**Handling of the *b=0***: Adds improvements to the manipulation of
the ``S0`` argument fed into DIPY models.