ENH: Allow `advanced_clip` to be performed in-place
Allow `advanced_clip` to be performed in-place: make the necessary
changes to the implementation so that operations (e.g. the `NumPy`
clipping, subtraction and rounding are performed in-place).

Adapt the call in the `ants` module to avoid static type checking
warnings.

Add a unit test for the `advanced_clip` function.