enh: improve general handling of parallelization
Brings improvements in parallelization management from #142, so they are
kept even if we finally decided against #142.

In particular, it opens up the implementation to set the ``step``
feature of DTI models.

It also extends the base data object with two convenience properties for
retrieving the 3D shape of the data and the number of voxels per volume.

X-References: #142.