DOC: Use `BibTeX` for references
Use `BibTeX` for references: adopt the `sphinxcontrib-bibtex` extension
to manage references.

Add a `referenves.bib` file to host the references, and a
`bibliography.rst` file to host them in the rendered documentation.

Cite the references in the docstrings using the footnote style. Adapt
the short descriptions where necessary since they cannot hold references
when using `sphinxcontrib-bibtex`.