BUG: Do not rely on URLs and use paths through S3
Do not rely on URLs and use paths through S3: some `urls` fields contain
multiple items, and thus, the assertion fails. Rather than picking one
of them (and even if the assumption that all URLs are the same or point
to the same resource), use S3, as it is experimentally verified that it
is twice as fast at least when retrieving the contents.