REF: Remove duplicate PET simulation data generation script
Remove duplicate PET simulation data generation script.

The script to simulate motion on a PET dataset was introduced in commit
f466d28 as part of the efforts to make PR
https://github.com/nipreps/nifreeze/pull/88
easier to merge.

The removed script was inadvertently re-introduced in commit f0d38de as
the works to solve the conflicts in
https://github.com/nipreps/nifreeze/pull/88 were resumed.