REF: Refactor PET estimator
Refactor PET estimator so that it is more closely aligned with the
`Estimator` class, and in order to eventually merge them.

Most notably, avoid creating train, test splits inside the estimator and
delegate the task of walking the sequence to the existing iterators.
This removes the need for the `lofo_split` method.