REF: Refactor PET model
Refactor PET model: use a base class that contains the essential
properties for a PET model and create a derived class that implements
the B-Spline correction.

In the base class:
- Remove the `timepoints` attribute: it corresponds to the PET dataset
  `midframe` attribute, and thus, there is no need to duplicate it in
  the model.
- Remove the `xlim` attribute: it corresponds to the PET dataset
  `total_duration` attribute, and thus, there is no need to duplicate it
  in the model.
- Remove the checks related to the above attributes: checking the
  coherence of their values is not the model's responsibility, and
  should be done when instantiating the dataset.

Make the model use the frame index instead of the `midframe` value as
the argument of the `fit_predict` function. This increases makes the PET
model be consistent with the approach used in the code base as a
principle for generalization in the framework.

The index was also marked as an integer by the type annotations (as it
should be expected), whereas `midframe` values that were being provided
are typically floating point values.

Change the tests and the PET notebook accordingly.

In the derived, BSpline class:
- Remove the unnecessary explicit `float` casting around the number of
  control points: the `dtype="float32"` specifier creates a float array.
  Fixes:
  ```
  Expected type 'str | Buffer | SupportsFloat | SupportsIndex', got 'Literal[0] | None | {__eq__} | int' instead
  ```
  raised by the IDE.

Take advantage of the commit to use `np.asarray` to avoid extra copies
when not necessary.

Cast explicitly to `int` when getting the frame indices in the
estimator's `run` method to avoid type checking errors:
```
src/nifreeze/estimator.py:274: error:
 Argument 1 to "fit_predict" of "BSplinePETModel" has incompatible type
 "signedinteger[_32Bit | _64Bit]"; expected "int | None"  [arg-type]
```

raised for example at:
https://github.com/nipreps/nifreeze/actions/runs/20402438629/job/58626813972#step:8:35