TST: Use a local random generator in spike identification test
Use a local random generator in spike identification test: when using
the `random_number_generator` fixture, every time `conftest.py` is
modified (even if the seed to the named fixture remains unchanged),
the spike indices change, and thus the test fails since the expected
indices are hard-coded. Using a local random generator avoids this
issue.

The rest of the tests using random data do not rely on particular values
being generated, and thus changes to `conftest.py` do not have an effect
on them.

Related to
https://github.com/nipreps/nifreeze/pull/181#issuecomment-3111742251
https://github.com/nipreps/nifreeze/pull/160#issuecomment-2954147226