ENH: Generalize iterators to accept `bvals` or `uptake` as kwargs
Generalize iterators to accept `bvals` or `uptake` as keyword arguments.
The previous implementation of the linear, random and centralsym
iterators was only accepting `bvals`. This patch set allows these
iterators to work with PET data through the `uptake` argument.

Document the functions by explicitly assigning the docstring to the
`__doc__` property of each function so that the `SIZE_KEYS_DOC` can be
reused and to allow the examples be run by the doctring tests.