TST: Test iterators
Test iterators. Even if docstring tests cover the core fetures,
exceptions remain untested. This patch set adds unit tests that do test
the exceptions and the different use cases with the keyword arguments
that iterators accept.

Take advantage of the commit to define the keyword arguments and
exception messages as global variables so that the exceptions raised by
the iterators get checked.