STY: Annotate DWI optional data class parameters as such
Annotate DWI optional data class parameters as such: add `None` to the
type annotation of `bzero` and `eddy_xfms` attributes.

Ignore the `bzero` indexing static checking error when serializing the
data to NIfTI: the indexing operation is protected by a conditional
`no_bzero` statement, so at that point the class instance is bound to
have a non-`None` `bzero` attribute value.

Assert that `bzero` is not `None` in tests to avoid static type checking
errors.