STY: Disallow `None` `dataobj` values
By definition, `dataobj` cannot be `None`, so remove the `None` type
annotations for it, and remove the related checks for `None` values as
they are no longer necessary.

For example, the `size3d` attribute of the `BaseDataset` class returns
```
np.prod(self.dataobj.shape[:3])
```

which would raise an error if `dataobj` is `None`, as a `NoneType` does
not have the `shape` attribute.