STY: Fix `motion_affines` type annotation
Fix `motion_affines` type annotation: add `None` to its type annotation
as it can also take that value, e.g.
https://github.com/nipreps/nifreeze/blob/eebdceb0ead93abab6a68685a684ece92cdfc526/src/nifreeze/data/dmri.py#L242
https://github.com/nipreps/nifreeze/blob/eebdceb0ead93abab6a68685a684ece92cdfc526/src/nifreeze/data/pet.py#L146
https://github.com/nipreps/nifreeze/blob/eebdceb0ead93abab6a68685a684ece92cdfc526/test/test_data_dmri.py#L90

Fixes:
```
test/test_data_dmri.py:88: error:
 Incompatible types in assignment (expression has type "ndarray[Any, dtype[Any]] | None",
 variable has type "ndarray[Any, Any]")  [assignment]
```

raised in
https://github.com/nipreps/nifreeze/actions/runs/18478441375/job/52648138075?pr=273#step:8:34

When not provided in the `PET` data class, build it as an array instead
of as a list. Avoids:
```
src/nifreeze/data/pet.py:147: error:
 Incompatible types in assignment (expression has type "list[None]",
 variable has type "ndarray[Any, Any] | None")  [assignment]
src/nifreeze/data/pet.py:149: error:
 Unsupported target for indexed assignment ("ndarray[Any, Any] | None")  [index]
```

raised locally when adding `None` to the type annotation.

Take advantage of the commit to explicitly mark it as an array in its
docstring instead of a list to avoid confusion.