REF: Refactor the `get_extra` attribute to return an array or `None`
Refactor the `get_extra` attribute to return an array or `None` instead
of a tuple: the derivative classes can only return an array as the
values they return can only be arrays, and the base class must return
`None` to remain generic and as it does not store any extra property.

Thus, there is no need to unpack the extras in the base class
`__getitem__` method.