fix: disallow the estimator to iterate over mock data
For some reason, our tests are sometimes trying to execute the full
estimator loop even on mock data.

By mocking the iterator (which now returns an empty list in the
offending test), we can now call ``run()`` on the estimator and it
returns without iterating over the dataset.

While this does not address the fundamental issue of the test (which is
a bit unclear, as it is not immediate that testing whether the estimator
generates the right model through the factory is part of the estimators'
responsibilities), at least the ``run()`` member returns a lot earlier
(and what's critical, without attempting to run ANTs'
``antsRegistration``).

Resolves #309.