fix: remove test data generation from session spin up
Removes the `motion_data` fixture as it depended on nitransforms and a
particular feature (apply on several volumes) that has been very
unreliable since 24.0.0.

I thought it was fixed in 25 but it doesn't seem to be the case
completely.

Moving the issue there by pre-caching moved data (opposite direction of #315,
but it will do for now).

Resolves: #315.
X-References: #209.