REF: Make dMRI b0 handling implementation more robust
Make dMRI b0 handling more robust: the current implementation would set
the `DWI` data class to whatever value is provided in the `bzero`
argument of the constructor. This could contain multiple b0 volumes, or
data whose shape does not match the shape of the DWI volumes. The first
case could result in a failure when serializing the data (depending on
the flags provided to the `to_nifti` function) since there would be a
mismatch when trying to concatenate the b0 data and the non-null
gradient data hold by the DWI instance. The second case is one that
should not be allowed.

Similarly, users are not warned about the opinionated decisions that
NiFreeze does when:
- Multiple `b=0` volumes are contained in the provided primary DWI data
  object/gradients and the median is computed.
- The primary DWI data object/gradients contain `b=0` data and a `bzero`
  value is provided, leading to a redundancy or discrepancy that
  requires to be resolved. NiFreeze would use the provided `bzero` data
  and discard the `b=0` in the primary DWI data object.

This patch set refactors the b0 handling at initialization to provide a
consistent and predictable behavior:
- Raises an exception when the provided `bzero` data does not match the
  shape of the DWI volumes. Ensures that `bzero` is a 3D volume, with no
  additional axes.
- Warns users when multiple `b=0` volumes are contained in the provided
  primary DWI data object/gradients and tells them that the median value
  will be used as a reference.
- Warns users when `b=0` volumes are detected in the primary DWI
  object/gradients and a `bzero` value is provided. Tells them that the
  primary object `b=0` values are discarded, and the externally set
  `bzero` data are used to resolve the redundancy/discrepancy.
- Squeezes the `b=0` volumes extracted from the primary DWI object if a
  single volume exists. Thus, when a single volume is found, not attempt
  to compute the mean across the singleton dimension is made.

Add the necessary tests to check that the b0 data matches the above
behavior, and modify the existing tests parametrization to avoid the now
disallowed cases (i.e. multiple volumes in `bzero` data) and to raise or
filter warnings where appropriate.

Add the necessary explanations to the `data.dmri` module documentation.