FIX: Test the DTI and DKI correctness
Test the DTI and DKI correctness: test that the `NiFreeze` DTI and DKI
model predictions give exactly the same results as when directly
instatiating the corresponding DIPY models.

Fixes a bug in the `NiFreeze` DKI model fitting: the DKI model requires
the b0 data to be present in the gradients and data. The `check_multi_b`
DIPY function that tests whether the dataset is a multi-shell dataset
will otherwise tell that there are not enough b-values, the
instantiation of the `DiffusionKurtosisModel` raising an exception.