TST: Refactor the random DWI data fixture
Refactor the random DWI data fixture: create `n_gradients` non-null
gradient DWI volumes, and return the null gradient volumes in the
`b0_dataobj` object. Avoids returning null gradients both in the
`dwi_dataobj` object and the `b0_dataobj` object, which is misleading.

This allows to simplify tests where the multiple b0 warning of the DWI
data class is not relevant by providing the null gradients implicitly
through the `dataobj` and `gradients` parameters.

Also, reduce the number of null gradients required to the
`random_gtab_data` fixture to only 1 where the multiple b0 conditions
does not need to be tested.

Refactor the `dwi_data_to_nifti` helper function: make the b0 dataobj
optional so that it is used only when necessary. Simplifies tests.